% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhibitr.R
\docType{class}
\name{exhibitr-class}
\alias{$,exhibitr-method}
\alias{[,exhibitr,ANY,ANY-method}
\alias{[[,exhibitr-method}
\alias{exhibitr}
\alias{exhibitr-class}
\alias{initialize,exhibitr-method}
\title{An S4 class for Exhibit objects.}
\usage{
\S4method{initialize}{exhibitr}(.Object, ..., title = character(0),
  subtitle = character(0), caption = character(0), gens = list())

\S4method{[[}{exhibitr}(x, i, j = NULL, ...)

\S4method{[}{exhibitr,ANY,ANY}(x, i, j = NULL)

\S4method{$}{exhibitr}(x, name)
}
\description{
An S4 class for Exhibit objects.

Constructor wrapper for exhibitr
}
\section{Slots}{

\describe{
\item{\code{title}}{Title for exhibit}

\item{\code{subtitle}}{Subtitle for exhibit}

\item{\code{caption}}{Caption for exhibit}

\item{\code{gens}}{List of exhibit generators. Each generator must be a function that
returns an exhibit component (e.g., a plot).}
}}

