% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_plottable.R
\name{shiny_plottable}
\alias{shiny_plottable}
\title{Create plottable object from exhibit}
\usage{
shiny_plottable(exhibit, data = NULL)
}
\value{
plottable List of 2 components, \code{ui} and \code{server}
}
\description{
Create plottable object from exhibit
}

