% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-df-filter.R
\name{shinyDataFilter}
\alias{shinyDataFilter}
\title{Create Shiny component for filtering data}
\usage{
shinyDataFilter(data, ...)
}
\arguments{
\item{data}{Dataframe}
}
\description{
Create Shiny component for filtering data
}

