% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhibitr.R
\docType{methods}
\name{xb_gens}
\alias{xb_gens}
\alias{xb_gens,exhibitr-method}
\title{List exhibit components}
\usage{
xb_gens(object)

\S4method{xb_gens}{exhibitr}(object)
}
\arguments{
\item{object}{Object to list generators of}
}
\description{
List exhibit components
}

