% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhibitr.R
\docType{methods}
\name{xb_shiny_module}
\alias{xb_shiny_module}
\alias{xb_shiny_module,NULL-method}
\alias{xb_shiny_module,exhibitr-method}
\alias{xb_shiny_module,xbc_gen-method}
\title{Generate shiny module from exhibit components or objects}
\usage{
xb_shiny_module(object, ...)

\S4method{xb_shiny_module}{xbc_gen}(object, ...)

\S4method{xb_shiny_module}{`NULL`}(object, ...)

\S4method{xb_shiny_module}{exhibitr}(object, ..., elems = "plot")
}
\arguments{
\item{.inner}{List of functions to be applied to the component outputs.}
}
\description{
Generate shiny module from exhibit components or objects
}

