% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhibitr.R
\docType{methods}
\name{xb_tabs_module}
\alias{xb_tabs_module}
\alias{xb_tabs_module,exhibitr-method}
\title{Create shiny module with tabs.}
\usage{
xb_tabs_module(object, elems = "plot", ...)

\S4method{xb_tabs_module}{exhibitr}(object, elems = "plot", ...,
  titles = NULL)
}
\description{
Create shiny module with tabs.
}

