% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhibitr.R
\docType{methods}
\name{xb_ui}
\alias{xb_ui}
\alias{xb_ui,exhibitr-method}
\alias{xb_ui,xbc_gen-method}
\title{Generic function for generating ui components.}
\usage{
xb_ui(object, id, ...)

\S4method{xb_ui}{xbc_gen}(object, id, ...)

\S4method{xb_ui}{exhibitr}(object, id, ..., elems = NULL)
}
\description{
Generic function for generating ui components.
}

