% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhibitr.R
\docType{class}
\name{xbc_datatable-class}
\alias{xbc_datatable}
\alias{xbc_datatable-class}
\title{S4 Class for datatable component generator}
\description{
S4 Class for datatable component generator

Constructor for xbc_datatable
}

