% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhibitr.R
\docType{class}
\name{xbc_gen-class}
\alias{xbc_gen}
\alias{xbc_gen-class}
\title{Class definition for exhibit component generator}
\description{
Class definition for exhibit component generator

Contructor function for xbc_gen
}

