% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhibitr.R
\docType{class}
\name{xbc_highchart-class}
\alias{xbc_highchart}
\alias{xbc_highchart-class}
\title{S4 class for highchart component generator}
\description{
S4 class for highchart component generator

Constructor for xbc_highchart
}

